fix bugs ex9
