+++ index.html
@@ -259,6 +259,9 @@
     </div>
   </template>
 
+  <!-- Сообщение об ошибке загрузки данных с сервера -->
+  <template id="error-downloading"></template>
+
   <script defer src="./leaflet/leaflet.js"></script>
   <script defer type="module" src="./js/main.js"></script>
 </body>
