+++ js/utils/requests.js
@@ -1,0 +1,22 @@
+export const getData = (onSuccess, onError) => () => {
+  fetch('https://23.javascript.pages.academy/keksobooking/data', {
+    method: 'GET',
+    credentials: 'same-origin',
+  }).then((response) => {
+    if (response.ok) {
+      return response.json();
+    }
+    throw new Error(`${response.status} ${response.statusText}`);
+  })
+    .then((json) => {
+      onSuccess(json);
+    })
+    .catch((err) => {
+      onError(err);
+    });
+};
+
+// Отправка данных - (задание 11 пункт 1)
+// export const postData = () => {
+
+// };
